ruby-init.sh supports rbenv-localized ruby

+++ scripts/ruby-init.sh
@@ -1,13 +1,15 @@
 #!/bin/sh
-cd `dirname $0`
 
-echo ---------- ruby, gem versions ----------
-ruby -v
-gem -v
+echo ------------- versions -------------
+echo "ruby: `ruby -v`"
+echo "gem:  `gem -v`"
 echo ---------- initializations ----------
+if [ -f Gemfile ]; then echo 'Gemfile detected, abort.'; exit 1; fi
+if [ -f Gemfile.lock ]; then echo 'Gemfile.lock detected, abort.'; exit 1; fi
+set -x
 gem update --system
-gem install bundler
-cp ../Gemfile .
+gem install -f bundler
+cp `dirname $0`/../Gemfile .
 bundle install
-bitclust setup
 rm Gemfile*
+if which rbenv >& /dev/null; then rbenv rehash ; fi
